WIP #1559 refactored blob for multiple blobs (w/ different backends)
